Separate users table component
